修复类型转换错误：Integer与Long互转
- UserFolderDao: 统一使用Integer参数类型
- UserFolderServiceImpl: 添加(long)id类型转换
- FileController: 处理JSON数字可能为Integer或Long的情况
- JettyStarter: 添加TLS 1.1/1.2支持和HTTP/1.1连接器

+++ src/main/java/love/linyi/dao/UserFolderDao.java
@@ -29,7 +29,7 @@ public interface UserFolderDao {
     void updateFileName(@Param("Path") String Path, @Param("id") int id);
 
     @Select("SELECT id, path, name, type FROM folder WHERE id = #{id} AND user_id = #{userId}")
-    UserFolder getFolderByIdAndUserId(@Param("id") Long id, @Param("userId") int userId);
+    UserFolder getFolderByIdAndUserId(@Param("id") Integer id, @Param("userId") int userId);
 
     @Select("SELECT id, path, name, type FROM folder WHERE id = #{id}")
     UserFolder getFolderById(@Param("id") Integer id);
@@ -38,7 +38,7 @@ public interface UserFolderDao {
     int countByPathAndUserId(@Param("path") String path, @Param("userId") int userId);
 
     @Update("UPDATE folder SET name = #{newName}, path = #{newPath} WHERE id = #{id} AND user_id = #{userId}")
-    int updateFolderNameAndPath(@Param("id") Long id, @Param("newName") String newName, @Param("newPath") String newPath, @Param("userId") int userId);
+    int updateFolderNameAndPath(@Param("id") Integer id, @Param("newName") String newName, @Param("newPath") String newPath, @Param("userId") int userId);
 
     @Update("UPDATE folder SET path = REPLACE(path, #{oldPath}, #{newPath}) WHERE user_id = #{userId} AND path LIKE CONCAT(#{oldPath}, '/%')")
     int updateChildrenPaths(@Param("oldPath") String oldPath, @Param("newPath") String newPath, @Param("userId") int userId);
